requirements.txt
Pillow==9.2.0
playsound==1.3.0
plyer==2.0.0
